blog-graphql-server/requirements.txt
aiodataloader==0.1.2
aiodns==1.1.1
aiohttp==3.4.4
aiohttp-graphql==1.0.0
aiohttp-route-middleware==0.0.2
async-timeout==3.0.0
asyncpg==0.17.0
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.4
cchardet==2.1.4
cffi==1.11.5
chardet==3.0.4
easydict==1.8
graphql-core==2.1
graphql-server-core==1.1.1
idna==2.7
idna-ssl==1.1.0
more-itertools==4.3.0
motor==2.0.0
motorodm==0.0.1
multidict==4.4.2
pluggy==0.7.1
promise==2.1
py==1.6.0
pycares==2.3.0
pycparser==2.19
PyJWT==1.6.4
pymongo==3.7.1
pytest==3.8.1
pytest-asyncio==0.9.0
pytest-runner==4.2
PyYAML==3.13
Rx==1.6.1
six==1.11.0
slugify==0.0.1
stringcase==1.2.0
typing==3.6.6
yarl==1.2.6
